+++ CHANGES.rst
@@ -8,6 +8,13 @@ Changelog
   Inline comments are invalid since editorconfig v0.15.0.
   [thet]
 
+- Refactor all tests to adapt to the new package structure created by cookieplone
+  [erral]
+
+- Use cookieplone to create an addon
+  [erral]
+
+
 
 7.0.0a3 (2025-07-22)
 --------------------